# trireg-java-client-api/target/generated-sources/openapi/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: example@example.com
    name: Trifork A/S
    url: https://github.com/trifork/trireg2-api
  description: An OpenAPI specification for time registrations to be used by a user
  license:
    name: MIT License
    url: https://github.com/trifork/trireg2-api/LICENSE.md
  title: TriReg2 User API
  version: 0.2.0
servers:
- description: Test server (uses test data)
  url: https://trireg2.tcs.trifork.dev
- description: Local development server
  url: http://0.0.0.0:8080
security:
- openId: []
- basicAuth: []
...
paths:
...
  /tag:
    delete:
      description: Deletes a tag that was added to a time registration
      operationId: deleteTagRegistration
      parameters:
      - description: The ID of the tag to delete
        explode: true
        in: query
        name: tagId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultDeleteResponse'
          description: Tag was successfully deleted
        "401":
          description: JWT is missing or invalid
        "404":
          description: The requested tag was not found and could not be deleted
      tags:
      - Tag
      x-accepts: application/json
    get:
      description: |
        Gets the tags associated with the specified time registration. This includes both the tags that have already been
        added, and the ones that can be added.
      operationId: getTimeRegistrationTags
      parameters:
      - description: The ID of the time registration to get tags for
        explode: true
        in: query
        name: timeRegistrationId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TimeRegistrationTag'
                type: array
          description: "List of tags for the requested time registration, both existing\
            \ and available to add"
        "401":
          description: JWT is missing or invalid
        "404":
          description: Unable to find a time registration with the specified timeRegistrationId
        "502":
          description: "While acting as a gateway or proxy, this service received\
            \ an invalid response from the upstream server."
      tags:
      - Tag
      x-accepts: application/json
    post:
      description: Tags a time registration with a specified tag and tag value
      operationId: tagTimeRegistration
      parameters:
      - description: The ID of the time registration to add a tag to
        explode: true
        in: query
        name: timeRegistrationId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagTimeRegistrationRequest'
        description: A JSON object containing tag information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultCreateResponse'
          description: Tag was successfully added to the time registration. The response
            contains the id of the new tag
        "401":
          description: JWT is missing or invalid
        "400":
          description: Tag was not added to the time registration due to an error
            with the request
        "404":
          description: No time registration was found to tag
        "502":
          description: "While acting as a gateway or proxy, this service received\
            \ an invalid response from the upstream server."
      tags:
      - Tag
      x-content-type: application/json
      x-accepts: application/json
    put:
      description: Updates a tag that was added to a time registration
      operationId: updateTagRegistration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRegistration'
        description: A JSON object containing the updated tag information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultUpdateResponse'
          description: Tag was successfully updated
        "400":
          description: The updated tag information yields an invalid tag
        "401":
          description: JWT is missing or invalid
        "404":
          description: The specified tag was not found
      tags:
      - Tag
      x-content-type: application/json
      x-accepts: application/json
...
components:
...
  schemas:
...
    DefaultCreateResponse:
      description: A response to a create operation.
      example:
        id: 12345
        status: The data was created successfully
      properties:
        id:
          description: The id of the created entity.
          format: int64
          type: integer
        status:
          description: A description of the status of the create operation.
          type: string
      required:
      - id
      - status
      type: object
...
    DefaultDeleteResponse:
      description: A response to a delete operation.
      example:
        id: 12345
        status: The data was deleted successfully
      properties:
        id:
          format: int64
          type: integer
        status:
          description: A description of the status of the delete operation.
          type: string
      required:
      - id
      - status
      type: object
...
    DefaultUpdateResponse:
      description: A response to an update operation.
      example:
        id: 12345
        status: The data was updated successfully
      properties:
        id:
          description: The id of the updated entity.
          format: int64
          type: integer
        status:
          description: A description of the status of the update operation.
          type: string
      required:
      - id
      - status
      type: object
...
    TimeRegistrationTag:
      description: "Composite object with information about a potential tag for a\
        \ time registration. This consists both of tags that has\nbeen set, and tags\
        \ that can potentially be set. Contains metadata describing the tag and the\
        \ current tag value (if \nany). The 'tagValue' is null if no tag value has\
        \ been set for the specific configuration.\n"
      example:
        tagConfigurationMetadata:
          tagConfigurationDescription: tagConfigurationDescription
          tagConfigurationId: 5
          tagConfigurationName: tagConfigurationName
          valueType: STRING
          cardinality: OPTIONAL
        tagRegistration:
          tagId: 2
          tagRegistration: tagRegistration
      properties:
        tagConfigurationMetadata:
          $ref: '#/components/schemas/TagConfigurationMetadata'
        tagRegistration:
          $ref: '#/components/schemas/TagRegistration'
      required:
      - tagConfigurationMetadata
      type: object
    TagRegistration:
      description: An existing tag with some value
      example:
        tagId: 2
        tagValue: tagValue
      properties:
        tagId:
          description: The id of this tag
          format: int64
          type: integer
        tagValue:
          description: "The value of this tag. This is a valid value corresponding\
            \ to the tag configuration. For example, if the tag \nconfiguration specifies\
            \ that the tag is an integer, then the value is a valid integer.\n"
          type: string
      required:
      - tagId
      - tagValue
      type: object
    TagTimeRegistrationRequest:
      description: |
        A request to set a tag value for a tag configuration for a time registration. Administrators configure the rules for
        tags for time registrations. This includes the format of the tag value, and whether tags are mandatory or optional.
        When setting a tag value for a time registration, the request must follow these rules.
      example:
        tagConfigurationId: 6
        value: value
      properties:
        tagConfigurationId:
          description: "The id of the tag configuration to set a value for. This must\
            \ be a valid tag configuration for the time \nregistration.\n"
          format: int64
          type: integer
        value:
          description: |-
            The value to set for the tag configuration. This is a string, but the format of the string is dependent on the tag
            configuration. For example, if the configuration specifies that the tag is an integer, then the string value must
            be a valid integer.
          type: string
      required:
      - tagConfigurationId
      - value
      type: object
...
    TagConfigurationMetadata:
      description: |
        Metadata for a tag configuration in the context of a specific time registration. This can differ based on the task the
        time registration is associated to, so a tag that is optional for one time registration can be mandatory for another.
      example:
        tagConfigurationDescription: tagConfigurationDescription
        tagConfigurationId: 5
        tagConfigurationName: tagConfigurationName
        valueType: STRING
        cardinality: OPTIONAL
      properties:
        tagConfigurationId:
          description: An identifier for this tag configuration
          format: int64
          type: integer
        tagConfigurationName:
          description: The name of the tag configuration
          type: string
        tagConfigurationDescription:
          description: An optional description of the tag configuration
          type: string
        cardinality:
          description: |
            The status of a time registration Valid values: * `OPTIONAL` - The tag can be added to this time registration, but it is not required. * `MANDATORY` - The tag must be added to this time registration. * `MANAGED` - The tag is managed by administrator configuration. It is read-only and can't be changed.
          enum:
          - OPTIONAL
          - MANDATORY
          - MANAGED
          type: string
        valueType:
          description: "The value type of this configuration Valid values: * `STRING`\
            \ - The tag value can be any string * `BOOL` - The tag value must be a\
            \ boolean on the form 'true' or 'false' (case insensitive) * `INT` - The\
            \ tag value must be a valid 32-bit integer"
          enum:
          - STRING
          - BOOL
          - INT
          type: string
      required:
      - cardinality
      - tagConfigurationId
      - tagConfigurationName
      - valueType
      type: object
...
  securitySchemes:
    openId:
      openIdConnectUrl: https://oidc.hosted.trifork.com/auth/realms/trifork/.well-known/openid-configuration
      type: openIdConnect
    basicAuth:
      scheme: basic
      type: http