# trireg-spring-server-api/target/generated-sources/openapi/src/main/java/com/trifork/trireg/server/api/TagApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.trifork.trireg.server.api;

import com.trifork.trireg.server.model.DefaultCreateResponse;
import com.trifork.trireg.server.model.DefaultDeleteResponse;
import com.trifork.trireg.server.model.DefaultUpdateResponse;
import com.trifork.trireg.server.model.TagRegistration;
import com.trifork.trireg.server.model.TagTimeRegistrationRequest;
import com.trifork.trireg.server.model.TimeRegistrationTag;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-08-21T22:43:49.173986300+02:00[Europe/Copenhagen]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Tag", description = "Tag for CRUD operations related to adding tags to time registrations")
public interface TagApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /tag
     * Deletes a tag that was added to a time registration
     *
     * @param tagId The ID of the tag to delete (required)
     * @return Tag was successfully deleted (status code 200)
     *         or JWT is missing or invalid (status code 401)
     *         or The requested tag was not found and could not be deleted (status code 404)
     */
    @Operation(
        operationId = "deleteTagRegistration",
        description = "Deletes a tag that was added to a time registration",
        tags = { "Tag" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tag was successfully deleted", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultDeleteResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "JWT is missing or invalid"),
            @ApiResponse(responseCode = "404", description = "The requested tag was not found and could not be deleted")
        },
        security = {
            @SecurityRequirement(name = "openId"),
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/tag",
        produces = { "application/json" }
    )
    
    default ResponseEntity<DefaultDeleteResponse> deleteTagRegistration(
        @NotNull @Parameter(name = "tagId", description = "The ID of the tag to delete", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "tagId", required = true) Long tagId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 12345, \"status\" : \"The data was deleted successfully\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /tag
     * Gets the tags associated with the specified time registration. This includes both the tags that have already been added, and the ones that can be added. 
     *
     * @param timeRegistrationId The ID of the time registration to get tags for (required)
     * @return List of tags for the requested time registration, both existing and available to add (status code 200)
     *         or JWT is missing or invalid (status code 401)
     *         or Unable to find a time registration with the specified timeRegistrationId (status code 404)
     *         or While acting as a gateway or proxy, this service received an invalid response from the upstream server. (status code 502)
     */
    @Operation(
        operationId = "getTimeRegistrationTags",
        description = "Gets the tags associated with the specified time registration. This includes both the tags that have already been added, and the ones that can be added. ",
        tags = { "Tag" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of tags for the requested time registration, both existing and available to add", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TimeRegistrationTag.class)))
            }),
            @ApiResponse(responseCode = "401", description = "JWT is missing or invalid"),
            @ApiResponse(responseCode = "404", description = "Unable to find a time registration with the specified timeRegistrationId"),
            @ApiResponse(responseCode = "502", description = "While acting as a gateway or proxy, this service received an invalid response from the upstream server.")
        },
        security = {
            @SecurityRequirement(name = "openId"),
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/tag",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<TimeRegistrationTag>> getTimeRegistrationTags(
        @NotNull @Parameter(name = "timeRegistrationId", description = "The ID of the time registration to get tags for", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "timeRegistrationId", required = true) Long timeRegistrationId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"tagConfigurationMetadata\" : { \"tagConfigurationDescription\" : \"tagConfigurationDescription\", \"tagConfigurationId\" : 5, \"tagConfigurationName\" : \"tagConfigurationName\", \"valueType\" : \"STRING\", \"cardinality\" : \"OPTIONAL\" }, \"tagRegistration\" : { \"tagId\" : 2, \"tagRegistration\" : \"tagRegistration\" } }, { \"tagConfigurationMetadata\" : { \"tagConfigurationDescription\" : \"tagConfigurationDescription\", \"tagConfigurationId\" : 5, \"tagConfigurationName\" : \"tagConfigurationName\", \"valueType\" : \"STRING\", \"cardinality\" : \"OPTIONAL\" }, \"tagRegistration\" : { \"tagId\" : 2, \"tagRegistration\" : \"tagRegistration\" } } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /tag
     * Tags a time registration with a specified tag and tag value
     *
     * @param timeRegistrationId The ID of the time registration to add a tag to (required)
     * @param tagTimeRegistrationRequest A JSON object containing tag information (required)
     * @return Tag was successfully added to the time registration. The response contains the id of the new tag (status code 200)
     *         or JWT is missing or invalid (status code 401)
     *         or Tag was not added to the time registration due to an error with the request (status code 400)
     *         or No time registration was found to tag (status code 404)
     *         or While acting as a gateway or proxy, this service received an invalid response from the upstream server. (status code 502)
     */
    @Operation(
        operationId = "tagTimeRegistration",
        description = "Tags a time registration with a specified tag and tag value",
        tags = { "Tag" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tag was successfully added to the time registration. The response contains the id of the new tag", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultCreateResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "JWT is missing or invalid"),
            @ApiResponse(responseCode = "400", description = "Tag was not added to the time registration due to an error with the request"),
            @ApiResponse(responseCode = "404", description = "No time registration was found to tag"),
            @ApiResponse(responseCode = "502", description = "While acting as a gateway or proxy, this service received an invalid response from the upstream server.")
        },
        security = {
            @SecurityRequirement(name = "openId"),
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/tag",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<DefaultCreateResponse> tagTimeRegistration(
        @NotNull @Parameter(name = "timeRegistrationId", description = "The ID of the time registration to add a tag to", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "timeRegistrationId", required = true) Long timeRegistrationId,
        @Parameter(name = "TagTimeRegistrationRequest", description = "A JSON object containing tag information", required = true) @Valid @RequestBody TagTimeRegistrationRequest tagTimeRegistrationRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 12345, \"status\" : \"The data was created successfully\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /tag
     * Updates a tag that was added to a time registration
     *
     * @param tagRegistration A JSON object containing the updated tag information (required)
     * @return Tag was successfully updated (status code 200)
     *         or The updated tag information yields an invalid tag (status code 400)
     *         or JWT is missing or invalid (status code 401)
     *         or The specified tag was not found (status code 404)
     */
    @Operation(
        operationId = "updateTagRegistration",
        description = "Updates a tag that was added to a time registration",
        tags = { "Tag" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tag was successfully updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultUpdateResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "The updated tag information yields an invalid tag"),
            @ApiResponse(responseCode = "401", description = "JWT is missing or invalid"),
            @ApiResponse(responseCode = "404", description = "The specified tag was not found")
        },
        security = {
            @SecurityRequirement(name = "openId"),
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/tag",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<DefaultUpdateResponse> updateTagRegistration(
        @Parameter(name = "TagRegistration", description = "A JSON object containing the updated tag information", required = true) @Valid @RequestBody TagRegistration tagRegistration
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 12345, \"status\" : \"The data was updated successfully\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}